add tests for module deps, and mov emodules till it passes
