+++ modules_basic/emoji.js
