+++ modules_core/index.test.js
@@ -1,0 +1,10 @@
+const test = require('tape')
+const combine = require('depject')
+
+const core = require('./')
+
+test('modules_core has no outside deps', t => {
+  t.ok(combine(core))
+  t.end()
+})
+
