+++ package.json
@@ -7,12 +7,20 @@
     "type": "git",
     "url": "git://github.com/dominictarr/patchbay.git"
   },
+  "scripts": {
+    "lite": "mkdir -p build && browserify index.js | indexhtmlify --title patchbay > build/index.html",
+    "start": "electro index.js",
+    "bundle": "mkdir -p build && browselectrify index.js > build/bundle.js",
+    "rebuild": "npm rebuild --runtime=electron --target=$(electron -v) --abi=$(electron --abi) --disturl=https://atom.io/download/atom-shell",
+    "graph": "node index.js | dot -Tsvg > graph.svg",
+    "test": "browserify modules_*/index.test.js | tape-run"
+  },
   "dependencies": {
     "@mmckegg/mutant": "^3.12.0",
     "brfs": "^1.4.3",
     "cont": "^1.0.3",
     "dataurl-": "^0.1.0",
-    "depject": "^3.1.0",
+    "depject": "^3.1.4",
     "es2040": "^1.2.4",
     "hjson": "^2.0.3",
     "human-time": "0.0.1",
@@ -65,21 +73,15 @@
     "browselectrify": "^1.0.1",
     "electro": "^2.0.3",
     "electron": "^1.4.10",
-    "indexhtmlify": "^1.3.1"
+    "indexhtmlify": "^1.3.1",
+    "tape": "^4.6.3",
+    "tape-run": "^2.1.5"
   },
   "browserify": {
     "transform": [
       "brfs",
       "es2040"
     ]
-  },
-  "scripts": {
-    "lite": "mkdir -p build && browserify index.js | indexhtmlify --title patchbay > build/index.html",
-    "start": "electro index.js",
-    "bundle": "mkdir -p build && browselectrify index.js > build/bundle.js",
-    "rebuild": "npm rebuild --runtime=electron --target=$(electron -v) --abi=$(electron --abi) --disturl=https://atom.io/download/atom-shell",
-    "graph": "node index.js | dot -Tsvg > graph.svg",
-    "test": "set -e; for t in test/*.js; do node $t; done"
   },
   "author": "Dominic Tarr <dominic.tarr@gmail.com> (http://dominictarr.com)",
   "license": "MIT"
